requirements.txt
requests
Flask
pycryptodome
python-dotenv
peewee
gunicorn
pymysqldocker